Fix type cast in image pixel read check
Corrects the type casting in the conditional that checks if the number of bytes read for image pixels matches the expected size.

+++ src/RageSurfaceUtils.cpp
@@ -809,7 +809,7 @@ RageSurface *RageSurfaceUtils::LoadSurface( RString file )
 		return nullptr;
 	}
 
-	if( f.Read( img->pixels, static_cast<size_t>(h.height) * h.pitch ) != static_cast<size_t>(h.height) * h.pitch )
+	if( static_cast<size_t>(f.Read( img->pixels, static_cast<size_t>(h.height) * h.pitch )) != static_cast<size_t>(h.height) * h.pitch )
 	{
 		delete img;
 		return nullptr;
